Changed accountId type to Long, added a checkAuthentication in AccountServer to check authTokens and get accountId back to be used as foreign keys etc.

+++ src/main/java/com/github/malow/accountserver/account/Account.java
@@ -3,7 +3,7 @@ package com.github.malow.accountserver.account;
 public class Account
 {
   // Persisted in database
-  public Integer id;
+  public Long id;
   public String username;
   public String password;
   public String email;
@@ -18,7 +18,7 @@ public class Account
 
   }
 
-  public Account(Integer id, String username, String password, String email, String pwResetToken, Integer failedLoginAttempts, String authToken)
+  public Account(Long id, String username, String password, String email, String pwResetToken, Integer failedLoginAttempts, String authToken)
   {
     this.id = id;
     this.username = username;
